pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytest-api-coverage"
version = "0.1.0"
description = "Pytest plugin for API test coverage analysis"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = [ "LICENSE" ]
keywords = [ "pytest", "pytest-xdist", "test", "testing", "api", "coverage", "swagger", "openapi"]
authors = [{ name = "Anatoly Bobunov", email = "dev.bobunov@gmail.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Framework :: Pytest",
]
dependencies = [
    "pytest>=7.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "httpx>=0.24.0",
    "jinja2>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/anatolybobunov/pytest-api-coverage"
Repository = "https://github.com/anatolybobunov/pytest-api-coverage"
Documentation = "https://github.com/anatolybobunov/pytest-api-coverage#readme"
Changelog = "https://github.com/anatolybobunov/pytest-api-coverage/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
    "pytest-mock",
    "pytest-xdist",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.22.0",
    "responses>=0.25.8",
    "ruff",
    "mypy>=1.0.0",
    "types-requests>=2.32.4.20260107",
    "types-pyyaml>=6.0.12.20250915",
]

[project.entry-points.pytest11]
api_coverage = "pytest_api_coverage.plugin"

[tool.hatch.build.targets.wheel]
packages = ["src/pytest_api_coverage"]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
# HTML template contains long Jinja2 lines
"src/pytest_api_coverage/writers/html_writer.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src/pytest_api_coverage"]
